python-blacksheep-psyconpg/requirements.txt
blacksheep
python-dotenv
asyncpg
uvicorn[standard]
black
